+++ src/main/java/com/foolsix/fancyenchantments/effect/Cumbersome.java
@@ -7,9 +7,11 @@ import net.minecraft.world.effect.MobEffectCategory;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 
-import static com.foolsix.fancyenchantments.effect.EffectIDs.CUMBERSOME_UUID;
+import java.util.UUID;
 
 public class Cumbersome extends MobEffect {
+    public static final String CUMBERSOME_NAME = "cumbersome";
+    public static final UUID CUMBERSOME_UUID = UUID.fromString("ce0565a4-c0e0-4a8b-902a-722516749f11");
     public static final ModConfig.CumbersomeOptions CONFIG = FancyEnchantments.getConfig().cumbersomeOptions;
 
     public Cumbersome() {